% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSTherm.R
\name{gen_profile}
\alias{gen_profile}
\title{Generate protein ratio profile from spectrum quantification matrix}
\usage{
gen_profile(x, method = "sum", method.denom = "first")
}
\arguments{
\item{x}{matrix of spectrum quantification values, one row per spectrum and
one column per channel}

\item{method}{method to use to "roll up" spectrum values to protein level}

\item{method.denom}{method used to determine ratio denominator, passed as
the "method" argument to \code{abs_to_ratio}}
}
\value{
A numeric vector of the same length as the number of matrix columns
}
\description{
\code{gen_profile} takes a matrix of spectrum channel
quantification values belonging to a protein and "rolls them up" into a
vector of protein-level relative quantification values
}

