% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSTherm.R
\name{model_genes}
\alias{model_genes}
\title{Model MSThermExperiment}
\usage{
model_genes(expt, genes, np, ...)
}
\arguments{
\item{expt}{An MSThermExperiment object}

\item{genes}{A vector of gene/protein IDs to model (default is all genes).}

\item{np}{Number of parallel jobs to start (default = number of available
processors)}

\item{min_rep_psm}{Minimum number of spectral matches required for each
replicate to model protein}

\item{min_smp_psm}{Minimum number of spectral matches required for each
sample to model protein}

\item{min_tot_psm}{Minimum number of spectral matches required across all
replicates to model protein}

\item{max_inf}{Maximum co-isolation interference level allowed to include a
spectrum in protein-level quantification}

\item{min_score}{minimum score allowed to include a
spectrum in protein-level quantification}

\item{max_score}{maximum score allowed to include a
spectrum in protein-level quantification}

\item{smooth}{(T/F) Perform loess smoothing on the data prior to modeling}

\item{method}{Protein quantification method to use (see Details)}

\item{method.denom}{Method used to calculate denominator of abundance
(see Details)}

\item{bootstrap}{(T/F) Perform bootstrap analysis to determine confidence
intervals (slow)}

\item{min_bs_psms}{Minimum number of spectral matches required to perform
bootstrapping}

\item{annot_sep}{Symbol used to separate protein group IDs (used for
retrieval of annotations) (default: '|')}
}
\value{
MSThermResultSet object
}
\description{
Model one or more proteins in an MSThermExperiment object
}
\details{
Valid quantification methods include:
  \describe{
     \item{"sum"}{Absolute sums for all spectra in each channel are used
        to calculate ratios}
     \item{"median"}{Median absolute intensities for spectra in each channel are used
        to calculate ratios}
     \item{"ratio.median"}{Ratios are calculated for each spectrum and
        the median ratio for each channel is used}
     \item{"ratio.mean"}{Ratios are calculated for each spectrum and
        the mean ratio for each channel is used}
  }
Valid denominator methods include:
  \describe{
     \item{"first"}{Use the value of the lowest temperature}
     \item{"max"}{Use the largest value}
     \item{"top3"}{Use the mean of the three largest values}
     \item{"near"}{Add description}
  }
}
\examples{
res  <- model_genes(expt,genes=c("gene01","gene02"))
res  <- model_genes(expt,np=4)

}

