% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstherm.R
\docType{package}
\name{mstherm}
\alias{mstherm}
\alias{mstherm-package}
\title{Model and analyze MS/MS-based protein melting data.}
\description{
\code{mstherm} is a package for modeling and analysis of
MS/MS-based thermal proteome profiling (TPP) experiments.
}
\author{
Jeremy Volkening \email{jdv@base2bio.com}
}
