% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize_to_std}
\alias{normalize_to_std}
\title{Normalize to a spike-in standard.}
\usage{
normalize_to_std(expt, protein)
}
\arguments{
\item{expt}{an MSThermExperiment object}

\item{protein}{ID of a protein to normalize against}
}
\value{
An MsThermExperiment object with normalized data slots
}
\description{
Normalizes each replicate of an experiment based on a given
spike-in protein standard (assumed to be present in equimolar amounts in
each channel).
}
\examples{
\dontrun{
expt <- normalize_to_std(expt, "cRAP_ALBU_BOVIN")
}

}

