% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSTherm.R
\name{normalize_to_tm}
\alias{normalize_to_tm}
\title{Re-normalize based on Tm}
\usage{
normalize_to_tm(expt, res)
}
\arguments{
\item{expt}{An MSThermExperiment object}

\item{res}{An MSThermResultSet object}
}
\value{
An MsThermExperiment object with re-normalized data slots
}
\description{
Normalizes each replicate of an experiment based on linear regression
of calculated Tm (corrects for remaining systematic error)
}
\examples{
\dontrun{
expt <- normalize_to_tm(expt, res)
}

}

