% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSTherm.R
\name{plot.MSThermResult}
\alias{plot.MSThermResult}
\title{Plot MSThermResult object}
\usage{
\method{plot}{MSThermResult}(result, table = T, col, CI.points = T,
  CI.Tm = T, ...)
}
\arguments{
\item{result}{An MSThermResult object}

\item{table}{(T/F) Include table of per-replicate parameters}

\item{col}{Array of colors used to plot samples}

\item{CI.points}{(T/F) Plot temperature point confidence intervals}

\item{CI.Tm}{(T/F) Plot Tm confidence intervals
intervals}
}
\value{
Nothing
}
\description{
Generate a denaturation plot for an modeled protein/group
}
\examples{
m  <- model_gene(expt,"P12345")
plot(m)

}

