% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSTherm.R
\name{plot.MSThermResultSet}
\alias{plot.MSThermResultSet}
\title{Plot MSThermResultSet object}
\usage{
\method{plot}{MSThermResultSet}(set, ...)
}
\arguments{
\item{set}{An MSThermResultSet object}

\item{...}{Other parameters are passed through to plot.MSThermResult}
}
\value{
Nothing
}
\description{
Generate a series of denaturation plots for all results in an MSThermResultSet
}
\details{
Since this function makes multiple sequential calls to
  plot.MSThermResult, it is usually used in conjunction with a multipage
  graphics device such as \code{"pdf()"}. Otherwise each subsequent call
  will only overwrite the previous output.
}
\examples{
res  <- model_genes(expt)
plot(res)

}

