% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSTherm.R
\name{summary.MSThermResult}
\alias{summary.MSThermResult}
\title{Summarize MSThermResult object}
\usage{
\method{summary}{MSThermResult}(result)
}
\arguments{
\item{result}{An MSThermResult object}
}
\value{
Nothing
}
\description{
Print a summary of an MSThermResult, including samples and parameters
}
\examples{
m  <- model_gene(expt,"P12345")
summmary(m)

}

