% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSTherm.R
\name{summary.MSThermResultSet}
\alias{summary.MSThermResultSet}
\title{Summarize MSThermResultSet object}
\usage{
\method{summary}{MSThermResultSet}(set)
}
\arguments{
\item{set}{An MSThermResultSet object}
}
\value{
Nothing
}
\description{
Print a summary of an MSThermResultSet, including samples and parameters
}
\examples{
res  <- model_genes(expt)
summmary(res)

}

