% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSTherm.R
\name{as.data.frame.MSThermResultSet}
\alias{as.data.frame.MSThermResultSet}
\title{MSResultSet to data frame}
\usage{
\method{as.data.frame}{MSThermResultSet}(x, ...)
}
\arguments{
\item{x}{an MSResultSet object}

\item{...}{additional arguments passed to or from other functions}
}
\value{
A dataframe populated with relevant information per result
}
\description{
Populates a dataframe with information from an MSResultSet, one row per
protein/group
}
\examples{
\dontrun{
df <- as.data.frame(expt)
write.table(df, "results.tsv")
}

}

