% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize_to_profile}
\alias{normalize_to_profile}
\title{Normalize to a profile.}
\usage{
normalize_to_profile(replicate, profile, model = T)
}
\arguments{
\item{replicate}{An MSThermReplicate object}

\item{profile}{A vector of relative values}

\item{model}{Whether to fit scale factors to model}
}
\value{
An MsThermReplicate object with normalized data slots
}
\description{
Normalizes an MSThermReplicate based on a pre-determined
vector of relative abundances
}
\examples{
\dontrun{
expt$ <- normalize_to_profile(expt, concentrations, model=T)
}

}
