% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSTherm.R
\name{plot.MSThermResult}
\alias{plot.MSThermResult}
\title{Plot MSThermResult object.}
\usage{
\method{plot}{MSThermResult}(x, table = T, col, CI.points = T, CI.Tm = T,
  ...)
}
\arguments{
\item{x}{An MSThermResult object}

\item{table}{(T/f) include table of per-replicate parameters}

\item{col}{array of colors used to plot samples}

\item{CI.points}{(T/F) plot temperature point confidence intervals}

\item{CI.Tm}{(T/F) plot Tm confidence intervals}

\item{...}{other parameters passed through to plot()}
}
\value{
Nothing
}
\description{
Generate a denaturation plot for an modeled protein/group.
}
\examples{
\dontrun{
m  <- model_protein(expt,"P38707")
plot(m)
}

}

