% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{summary.MSThermResult}
\alias{summary.MSThermResult}
\title{Summarize MSThermResult object.}
\usage{
\method{summary}{MSThermResult}(object, ...)
}
\arguments{
\item{object}{an MSThermResult object}

\item{...}{additional arguments passed to or from other functions}
}
\value{
Nothing
}
\description{
Print a summary of an MSThermResult, including samples and
parameters.
}
\examples{
# see model_protein() for an example

}
