% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSTherm.R
\name{summary.MSThermResultSet}
\alias{summary.MSThermResultSet}
\title{Summarize MSThermResultSet object}
\usage{
\method{summary}{MSThermResultSet}(object, ...)
}
\arguments{
\item{object}{an MSThermResultSet object}

\item{...}{additional arguments passed to or from other functions}
}
\value{
Nothing
}
\description{
Print a summary of an MSThermResultSet, including samples and parameters
}
\examples{
\dontrun{
res  <- model_experiment(expt)
summmary(res)
}

}

